postgres_cdc: clean up tests
